+++ changelog.md
@@ -1,16 +1,23 @@
-o# Changelog
+# Changelog
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## 0.0.6 - 2018-00-01
+## 0.0.6 - 2018-09-01, 2018-09-03,
 ### Added
 - Motive content
 - Footer content
+- Database helpers
+- Post timestamp
 
 ### Changed
 - Artist card link colours
+- Home page rendering
+- Home page styling
+- Home page chronological order
+- Typography sizing
+
 
 ## 0.0.5 - 2018-08-29
 ### Added